Update mining speed monitor

+++ exccjson/chainsvrresults.go
@@ -461,7 +461,7 @@ type GetMiningInfoResult struct {
 	Errors           string  `json:"errors"`
 	Generate         bool    `json:"generate"`
 	GenProcLimit     int32   `json:"genproclimit"`
-	HashesPerSec     int64   `json:"hashespersec"`
+	HashesPerSec     float64 `json:"hashespersec"`
 	NetworkHashPS    int64   `json:"networkhashps"`
 	PooledTx         uint64  `json:"pooledtx"`
 	TestNet          bool    `json:"testnet"`
